DRY refactor for toErrorReport
