+++ src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/GoogleServicesDAO.scala
@@ -11,7 +11,8 @@ import org.joda.time.DateTime
 import spray.http.StatusCodes
 import scala.concurrent.Future
 
-abstract class GoogleServicesDAO(groupsPrefix: String) {
+abstract class GoogleServicesDAO(groupsPrefix: String) extends ErrorReportable {
+  override val errorReportSource = "google"
 
   // returns a workspaceID
   def setupWorkspace(userInfo: UserInfo, projectId: String, workspaceId: String, workspaceName: WorkspaceName): Future[GoogleWorkspaceInfo]
@@ -66,17 +67,6 @@ abstract class GoogleServicesDAO(groupsPrefix: String) {
 
   def toProxyFromUser(userSubjectId: RawlsUserSubjectId): String
   def toUserFromProxy(proxy: String): String
-
-  def toErrorReport(throwable: Throwable) = {
-    val SOURCE = "google"
-    throwable match {
-      case gjre: GoogleJsonResponseException =>
-        val statusCode = StatusCodes.getForKey(gjre.getStatusCode)
-        ErrorReport(SOURCE,ErrorReport.message(gjre),statusCode,ErrorReport.causes(gjre),Seq.empty)
-      case _ =>
-        ErrorReport(SOURCE,ErrorReport.message(throwable),None,ErrorReport.causes(throwable),throwable.getStackTrace)
-    }
-  }
 
   def getUserCredentials(rawlsUserRef: RawlsUserRef): Future[Option[Credential]]
 }
